model/DeOldify/colab_requirements.txt
fastai==1.0.51
tensorboardX==1.6
ffmpeg-python
youtube-dl>=2019.4.17
opencv-python>=3.3.0.10
pillow
tornado~=5.1.0
torch==1.0.1
